% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_citation.R
\name{create_citation}
\alias{create_citation}
\title{Create a biocthis-style CITATION template}
\usage{
create_citation()
}
\value{

}
\description{
This function is very similar to \code{usethis::use_citation()} except
that it uses a template from \code{biocthis}.
}
\examples{

\dontrun{
create_citation()
}

}
