% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_news_md.R
\name{create_news_md}
\alias{create_news_md}
\title{Create a biocthis-style news.md file}
\usage{
create_news_md(open = rlang::is_interactive())
}
\arguments{
\item{open}{Whether to open the resulting file.}
}
\value{

}
\description{
This function is very similar to \code{usethis::use_news_md()} except that it
uses a template from \code{biocthis}.
}
\details{
Unlike the original function, \code{create_news_md()} does not make git commits.
}
\examples{

 \dontrun{
create_news_md()
}

}
