% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_readme_rmd.R
\name{create_readme_rmd}
\alias{create_readme_rmd}
\title{Create a biocthis-style README.Rmd file}
\usage{
create_readme_rmd(open = rlang::is_interactive())
}
\arguments{
\item{open}{Whether to open the resulting file.}
}
\value{

}
\description{
This function is very similar to \code{usethis::use_readme_rmd()} except that it
uses a template from \code{biocthis}.
}
\examples{

 \dontrun{
create_readme_rmd()
}
}
