% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vignette.R
\name{create_vignette}
\alias{create_vignette}
\title{Create a biocthis-style vignette template}
\usage{
create_vignette(name, title = name)
}
\value{

}
\description{
This function is very similar to \code{usethis::use_vignette()} except
that it uses a template from \code{biocthis}.
}
\examples{

 \dontrun{
pkg <- basename(here::here())
biocthis::create_vignette(pkg, paste("Introduction to", pkg))
}

}
