% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_citation.R
\name{use_bioc_citation}
\alias{use_bioc_citation}
\title{Create a biocthis-style CITATION template}
\usage{
use_bioc_citation()
}
\value{
This function adds and/or replaces the \code{inst/CITATION} file in your
R package.
}
\description{
This function is very similar to \code{usethis::use_citation()} except
that it uses a template from \code{biocthis}. This template can help you get
started with a citation file for your Bioconductor package. You can
compare the template produced with the one from \code{usethis::use_citation()}
to pick and choose the parts you like best from each.
}
\examples{

\dontrun{
## Run this function in your package
use_bioc_citation()
}

## Here's an example with a temporary package

## Set the package name
pkgname <- "biocthisexample"

## Create the example package in a temporary location
withr::with_dir(tempdir(), {
    usethis::create_package(pkgname)
})

## Save the path to our temporary package for the rest of the examples
pkgdir <- file.path(tempdir(), pkgname)

## Create a template CITATION file that is Bioconductor-friendly
withr::with_dir(pkgdir, {
    biocthis::use_bioc_citation()
})
}
