% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_feature_request_template.R
\name{use_bioc_feature_request_template}
\alias{use_bioc_feature_request_template}
\title{Create a biocthis-style GitHub feature request template}
\usage{
use_bioc_feature_request_template()
}
\value{
This function adds and/or replaces the
\code{.github/ISSUE_TEMPLATE/feature_request_template.md} file in your R package.
}
\description{
This function is very similar to \code{usethis::use_tidy_issue_template()} except
that it uses a template from \code{biocthis} and is tailored for feature requests.
}
\details{
For more details on how this function came to be, check
\url{https://github.com/lcolladotor/biocthis/pull/33}.
}
\examples{
\dontrun{
## Run this function in your package
use_bioc_feature_request_template()
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg()

## Create a feature request template GitHub file that is Bioconductor-friendly
biocthis::use_bioc_feature_request_template()
}
