% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_issue.R
\name{use_bioc_issue_template}
\alias{use_bioc_issue_template}
\title{Create a biocthis-style GitHub issue template}
\usage{
use_bioc_issue_template()
}
\value{

}
\description{
This function is very similar to \code{usethis::use_tidy_issue_template()} except
that it uses a template from \code{biocthis}.
}
\examples{

 \dontrun{
use_bioc_issue_template()
}

}
