% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_pkgdown_css.R
\name{use_bioc_pkgdown_css}
\alias{use_bioc_pkgdown_css}
\title{Create a css with Bioconductor colors for pkgdown}
\usage{
use_bioc_pkgdown_css()
}
\value{
This function adds and/or replaces the \code{pkgdown/extra.css} file in
your R package.
}
\description{
This function creates the \code{pkgdown/extra.css} file with Bioconductor-style
colors that will make your \code{pkgdown} documentation websites much cooler ^_^.
}
\details{
For more details on how this function came to be, check
\url{https://github.com/lcolladotor/biocthis/issues/34}.
}
\examples{
\dontrun{
## Run this function in your package
use_bioc_support()
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg()

## Create a css file to style your pkgdown website in Bioconductor colors
biocthis::use_bioc_pkgdown_css()
}
