% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tinytest.R
\name{use_tinytest}
\alias{use_tinytest}
\alias{use_tiny_test}
\title{Use the \code{tinytest} framework for unit tests}
\usage{
use_tinytest()

use_tiny_test(name = NULL, open = interactive())
}
\arguments{
\item{name}{Either a name without extension, or \code{NULL} to create the
paired file based on currently open file in the script editor. If
the \verb{R/} file is open, \code{use_test()} will create/open the corresponding
test file; if the test file is open, \code{use_r()} will create/open the
corresponding \verb{R/} file.}

\item{open}{Whether to open the file for interactive editing.}
}
\value{
\itemize{
\item use_tinytest: creates a \code{tinytest.R} file in the \code{tests} directory
\item use_tiny_test: creates a \verb{test_<name>.R} file in the \code{inst/tinytest}
directory where are to reside
}
}
\description{
This function works similarly to \code{usethis::use_testthat}. It sets up the
minimal files required to make \code{tinytest} compatible with the current
project.
}
\section{Functions}{
\itemize{
\item \code{use_tiny_test()}: add a unit test file for testing functions in the
current R file

}}
