% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_citation.R
\name{use_bioc_citation}
\alias{use_bioc_citation}
\title{Create a biocthis-style CITATION template}
\usage{
use_bioc_citation()
}
\value{
This function adds and/or replaces the \code{inst/CITATION} file in your
R package.
}
\description{
This function is very similar to \code{usethis::use_citation()} except
that it uses a template from \code{biocthis}. This template can help you get
started with a citation file for your Bioconductor package. You can
compare the template produced with the one from \code{usethis::use_citation()}
to pick and choose the parts you like best from each.
}
\examples{

\dontrun{
## Run this function in your package
use_bioc_citation()
}

## Set the package name on a temporary directory
pkgdir <- file.path(tempdir(), "biocthisexample")

## Create the example package in a temporary location
usethis::create_package(pkgdir)

## Set a local project to work with for the examples
usethis::local_project(pkgdir)

## Create a template CITATION file that is Bioconductor-friendly
biocthis::use_bioc_citation()
}
