% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_citation.R
\name{use_bioc_citation}
\alias{use_bioc_citation}
\title{Create a biocthis-style CITATION template}
\usage{
use_bioc_citation()
}
\value{

}
\description{
This function is very similar to \code{usethis::use_citation()} except
that it uses a template from \code{biocthis}. This template can help you get
started with a citation file for your Bioconductor package. You can
compare the template produced with the one from \code{usethis::use_citation()}
to pick and choose the parts you like best from each.
}
\examples{

\dontrun{
use_bioc_citation()
}

}
