% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_description.R
\name{use_bioc_description}
\alias{use_bioc_description}
\title{Create a biocthis-style DESCRIPTION template}
\usage{
use_bioc_description(biocViews = "Software")
}
\arguments{
\item{biocViews}{A \code{character()} with the Bioconductor biocViews terms you
want to use for your package. See
\url{https://bioconductor.org/packages/release/BiocViews.html} for details. Note
that the terms you choose have to be part of one of the main four trees:
software, annotation, experiment or workflow.}
}
\value{
This function adds and/or replaces the \code{DESCRIPTION} file in your
R package.
}
\description{
This function is very similar to \code{usethis::use_description()} except
that it uses a template from \code{biocthis}. This template includes some of the
information expected in the DESCRIPTION file of a Bioconductor package. You
can compare the output of this function to the one from
\code{usethis::use_description()} to select the parts you want from each.
}
\details{
For details about the DESCRIPTION file guidelines for Bioconductor
packages check
\url{http://bioconductor.org/developers/package-guidelines/#description}.
}
\examples{

\dontrun{
## Run this function in your package
use_bioc_description()
}

## Here's an example with a temporary package

## Set the package name
pkgname <- "biocthisexample"

## Create the example package in a temporary location
withr::with_dir(tempdir(), {
    usethis::create_package(pkgname)
})

## Save the path to our temporary package for the rest of the examples
pkgdir <- file.path(tempdir(), pkgname)

## Setup up git
withr::with_dir(pkgdir, {
    usethis::use_git()
})

## Create a template DESCRIPTION file that is Bioconductor-friendly
withr::with_dir(pkgdir, {
    biocthis::use_bioc_description()
})
}
