% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_description.R
\name{use_bioc_description}
\alias{use_bioc_description}
\title{Create a biocthis-style DESCRIPTION template}
\usage{
use_bioc_description(biocViews = "Software")
}
\arguments{
\item{biocViews}{A \code{character()} with the Bioconductor biocViews terms you
want to use for your package. See
\url{https://bioconductor.org/packages/release/BiocViews.html} for details. Note
that the terms you choose have to be part of one of the main four trees:
software, annotation, experiment or workflow.}
}
\value{
This function adds and/or replaces the \code{DESCRIPTION} file in your
R package.
}
\description{
This function is very similar to \code{usethis::use_description()} except
that it uses a template from \code{biocthis}. This template includes some of the
information expected in the DESCRIPTION file of a Bioconductor package. You
can compare the output of this function to the one from
\code{usethis::use_description()} to select the parts you want from each.
}
\details{
For details about the DESCRIPTION file guidelines for Bioconductor
packages check
\url{http://bioconductor.org/developers/package-guidelines/#description}.
}
\examples{

\dontrun{
## Run this function in your package
use_bioc_description()
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg(use_git = TRUE)

## Create a template DESCRIPTION file that is Bioconductor-friendly
biocthis::use_bioc_description()
}
