% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_description.R
\name{use_bioc_description}
\alias{use_bioc_description}
\title{Create a biocthis-style DESCRIPTION template}
\usage{
use_bioc_description(biocViews = "Software")
}
\value{

}
\description{
This function is very similar to \code{usethis::use_description()} except
that it uses a template from \code{biocthis}. This template includes some of the
information expected in the DESCRIPTION file of a Bioconductor package. You
can compare the output of this function to the one from
\code{usethis::use_description()} to select the parts you want from each.
}
\details{
For details about the DESCRIPTION file guidelines for Bioconductor
packages check
\url{http://bioconductor.org/developers/package-guidelines/#description}.
}
\examples{

\dontrun{
use_bioc_description()
}

}
