% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_github_action.R
\name{use_bioc_github_action}
\alias{use_bioc_github_action}
\title{Create a biocthis-style GitHub action workflow}
\usage{
use_bioc_github_action()
}
\value{
This function adds and/or replaces the
\code{.github/workflows/check-bioc.yml} file in your R package.
}
\description{
This function is very similar to \code{usethis::use_github_action()} except
that it uses a template from \code{biocthis}. It creates a Bioconductor-friendly
GitHub action workflow for your package.
}
\details{
For the full history on how this GitHub actions workflow came to be, check
\url{https://github.com/r-lib/actions/issues/84}.
}
\examples{

\dontrun{
biocthis::use_bioc_github_action()
}

}
