% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_github_action.R
\name{use_bioc_github_action}
\alias{use_bioc_github_action}
\title{Create a biocthis-style GitHub Actions workflow}
\usage{
use_bioc_github_action()
}
\value{
This function adds and/or replaces the
\code{.github/workflows/check-bioc.yml} file in your R package.
}
\description{
This function is very similar to \code{usethis::use_github_action()} except
that it uses a template from \code{biocthis}. It creates a Bioconductor-friendly
GitHub action workflow for your package. You can also use this GitHub
Actions workflow by executing
\code{usethis::use_github_action("check-bioc", "https://bit.ly/biocthis_gha", "check-bioc.yml")}
without having to install \code{biocthis}.
}
\details{
For the full history on how this GitHub Actions workflow came to be, check
the "biocthis developer notes" vignette
\url{https://lcolladotor.github.io/biocthis/articles/biocthis_dev_notes.html}.
}
\examples{

\dontrun{
## Run this function in your package
biocthis::use_bioc_github_action()

## You can also use this GitHub Actions workflow without installing biocthis
## by running:
usethis::use_github_action(
    "check-bioc", "https://bit.ly/biocthis_gha",
    "check-bioc.yml"
)
}

## Create an example package for illustrative purposes.
## Note: you do not need to run this for your own package!
pkgdir <- biocthis_example_pkg()

## Create a GitHub Actions workflow that is Bioconductor-friendly
biocthis::use_bioc_github_action()
}
