% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_issue_template.R
\name{use_bioc_issue_template}
\alias{use_bioc_issue_template}
\title{Create a biocthis-style GitHub issue template}
\usage{
use_bioc_issue_template()
}
\value{
This function adds and/or replaces the
\code{.github/ISSUE_TEMPLATE/issue_tempalte.md} file in your R package.
}
\description{
This function is very similar to \code{usethis::use_tidy_issue_template()} except
that it uses a template from \code{biocthis}.
}
\details{
For more details on how this function came to be, check
\url{https://github.com/r-lib/usethis/issues/1108}.
}
\examples{

\dontrun{
## Run this function in your package
use_bioc_issue_template()
}

## Set the package name on a temporary directory
pkgdir <- file.path(tempdir(), "biocthisexample")

## Create the example package in a temporary location
usethis::create_package(pkgdir)

## Set the project to use for the examples
usethis::proj_set(pkgdir)

## Create a template GitHub issue file that is Bioconductor-friendly
biocthis::use_bioc_issue_template()
}
