% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_news_md.R
\name{use_bioc_news_md}
\alias{use_bioc_news_md}
\title{Create a biocthis-style news.md file}
\usage{
use_bioc_news_md(open = rlang::is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio,
if applicable, or via \code{utils::file.edit()} otherwise. This argument has the
same behavior as the \code{open} argument used in \code{usethis} functions.}
}
\value{

}
\description{
This function is very similar to \code{usethis::use_news_md()} except that it
uses a template from \code{biocthis}.
}
\details{
Unlike the original function, \code{use_bioc_news_md()} does not make git commits.
So you won't lose work you didn't save when using this function.
}
\examples{

 \dontrun{
use_bioc_news_md()
}

}
