% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_pkg_templates.R
\name{use_bioc_pkg_templates}
\alias{use_bioc_pkg_templates}
\title{Create a biocthis package template R scripts}
\usage{
use_bioc_pkg_templates(open = rlang::is_interactive())
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio,
if applicable, or via \code{utils::file.edit()} otherwise. This argument has the
same behavior as the \code{open} argument used in \code{usethis} functions.}
}
\value{

}
\description{
This function creates the \code{dev} directory and in it, it creates R script
files that you can follow to create your own Bioconductor-friendly R
package.
}
\examples{

 \dontrun{
use_bioc_pkg_templates()
}

}
