% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_support.R
\name{use_bioc_support}
\alias{use_bioc_support}
\title{Create a biocthis-style GitHub support file}
\usage{
use_bioc_support()
}
\value{
This function adds and/or replaces the \code{.github/SUPPORT.md} file in
your R package.
}
\description{
This function is very similar to \code{usethis::use_tidy_support()} except that it
uses a template from \code{biocthis}.
}
\details{
For more details on how this function came to be, check
\url{https://github.com/r-lib/usethis/issues/1108}.
}
\examples{

\dontrun{
## Run this function in your package
use_bioc_support()
}

## Here's an example with a temporary package

## Set the package name
pkgname <- "biocthisexample"

## Create the example package in a temporary location
withr::with_dir(tempdir(), {
    usethis::create_package(pkgname)
})

## Save the path to our temporary package for the rest of the examples
pkgdir <- file.path(tempdir(), pkgname)

## Create a template GitHub support file that is Bioconductor-friendly
withr::with_dir(pkgdir, {
    biocthis::use_bioc_support()
})
}
