% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_support.R
\name{use_bioc_support}
\alias{use_bioc_support}
\title{Create a biocthis-style GitHub support file}
\usage{
use_bioc_support()
}
\value{

}
\description{
This function is very similar to \code{usethis::use_tidy_support()} except that it
uses a template from \code{biocthis}.
}
\details{
For more details on how this function came to be, check
\url{https://github.com/r-lib/usethis/issues/1108}.
}
\examples{

\dontrun{
use_bioc_support()
}

}
