% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_bioc_vignette.R
\name{use_bioc_vignette}
\alias{use_bioc_vignette}
\title{Create a biocthis-style vignette template}
\usage{
use_bioc_vignette(name, title = name)
}
\value{

}
\description{
This function is very similar to \code{usethis::use_vignette()} except
that it uses a template from \code{biocthis}. This template includes
instructions for citing other packages using \code{knitcitations}, uses
\code{sessioninfo::session_info()} for displaying the R session information since
it includes GitHub installation information and other useful details. The
template also includes a section on how to ask for help and required
knowledge.
}
\examples{

 \dontrun{
pkg <- basename(here::here())
biocthis::use_bioc_vignette(pkg, paste("Introduction to", pkg))
}

}
